refactors and unifies requirements extractor
